+++ http4k-server-ratpack/src/main/kotlin/org/http4k/server/Ratpack.kt
@@ -5,7 +5,6 @@ import org.http4k.core.Method
 import org.http4k.core.Request
 import org.http4k.core.RequestSource
 import org.http4k.core.Response
-import org.http4k.core.Status
 import ratpack.handling.Context
 import ratpack.handling.Handler
 import ratpack.http.TypedData
@@ -14,7 +13,7 @@ import ratpack.server.RatpackServerSpec
 import ratpack.server.ServerConfig.builder
 
 class Ratpack(port: Int = 8000) : ServerConfig {
-    private val serverConfig = builder().port(port)
+    private val serverConfig = builder().connectQueueSize(1000).port(port)
 
     override fun toServer(httpHandler: HttpHandler): Http4kServer {
         val server = RatpackServer.of { server: RatpackServerSpec ->
@@ -61,9 +60,4 @@ class RatpackHttp4kHandler(private val httpHandler: HttpHandler) : Handler {
         context.response.status(status.code)
         context.response.send(body.payload.array())
     }
-}
-
-
-fun main() {
-    { req: Request -> Response(Status.OK).body("foobar") }.asServer(Ratpack(9000)).start()
 }